requirements.txt
# Hypixel Skyblock Scripts — Python Dependencies
# Install with: pip install -r requirements.txt

pyautogui>=0.9.54
pynput>=1.7.6
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pyinstaller>=6.0.0   # For building standalone .exe files
pytesseract>=0.3.10  # OCR for AH sniper and Bazaar flipper (also requires Tesseract binary)
